backend/pyproject.toml
[tool.poetry]
name = "flat-finder-backend"
version = "0.1.0"
description = "Backend services for Flat Finder POC"
authors = ["Flat Finder Team <dev@flatfinder.local>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
aiohttp = "^3.10.5"
selenium = "^4.24.0"
playwright = "^1.47.0"
pydantic = "^2.9.2"
supabase = "^2.5.1"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
